requirements.txt
PyQt5>=5.15.0
requests>=2.25.0
python-dateutil>=2.8.1
numpy>=1.19.0
pandas>=1.1.0
matplotlib>=3.3.0
scipy>=1.5.0
scikit-learn>=0.24.0
tensorflow>=2.4.0
python-dotenv>=1.1.0
xgboost>=3.0.2
aiohttp>=3.12.13
tqdm>=4.67.1
joblib>=1.0.0
catboost>=1.2.8
swifter>=1.4.0
tabulate>=0.8.9